Create 2021-08-07-Self-Worth-DailyBlogPost.md
---
layout: post
title: Self-Worth Discovery, It's a journey
category: blog
keywords: blog writing author shikhashikz writing writinglife writingcommunity dailyblogpost dailyblogpostchallenge happiness suffering life experiences 
image: assets/images/Post1281.jpg

---

What’s the whole drama on self-worth...

***My bonding with this term is beginning to firm up as I am exploring the concept of Personal Branding. This led me to find the literal meaning of Self-Worth.***

Dictionary.com defines self-worth as:

**the sense of one's value or worth as a person; self-esteem; self-respect.**

My attention kept on coming back to the word “self” when I divided this definition into three parts. I had a question that who will decide my self-worth and what am I capable of. The definition did make it easy for me.

The sense of one’s value or worth as a person: It’s I who have to figure out my value and worth as a person. It will be foolish to entrust this important aspect in another person’s hand to “Decide your self-worth”

But then it did lead me to the second question: How would I know my worth? Is it related to what kind of things I can accomplish wrt my skills/ expertise? Is it quantitative? If I do not have mastery in any of the subjects, does it mean that I am worthless? Does it increases or decreases with time? For me, I can firmly say that my self-worth has increased in the past decade. 

Coming to the second part which says self-esteem, Dictionary.com defines it as:

-realistic respect for or favourable impression of oneself; self-respect

-an inordinately or exaggeratedly favourable impression of oneself

Here as well, it talks about how we treat ourselves and reflecting on our behaviour and reaction to situations, what’s our impression in our eyes. This is a hard bit as we do turn blind eyes to our reactions and fail to hold ourselves accountable.

Yes, there is a dangerous side to it as well, when we inflate our ego and think HIGH of our self, we might get thrown into the muddy puddles. We all are the SUPERHERO of our movies. It’s ok to be the Hero, but let’s be realistic and don’t blow the balloon so hard that it bursts.

On the last part which is Self-respect, Dictionary.com defines it as:

proper esteem or regard for the dignity of one's character.

Mark the word, “proper” and “regard” 

I do equate self-worth with personal branding and my three observations are:

•	I have to discover my self-worth

•	I have to be a keen observer of my behaviour to establish a realistic self-esteem

•	I have to hold myself accountable for a dignified character

Hope, now like me, you are answerable for your self-worth and not somebody else.